chore: Import local jars through gradle
Have to do this if I want to run the project in IntelliJ. Adding the jars in a way similar to the way we do it in Eclipse doesn't work, I think because gradle overrides the added jars. In fact, if my testing was correct, this is preferable for Eclipse as well, since gradle rebuilds in Eclipse seem to nuke the external jars added to build path?